fix CI issues (iwyu, black, compile)

+++ lib/opencl/clDeviceInterface.cpp
@@ -51,11 +51,11 @@ void clDeviceInterface::free_gpu_memory(void* ptr) {
 }
 
 
-cl_mem clDeviceInterface::get_gpu_memory(const std::string& name, const uint32_t len) {
+cl_mem clDeviceInterface::get_gpu_memory(const std::string& name, const size_t len) {
     return (cl_mem)gpuDeviceInterface::get_gpu_memory(name, len);
 }
 cl_mem clDeviceInterface::get_gpu_memory_array(const std::string& name, const uint32_t index,
-                                               const uint32_t len) {
+                                               const size_t len) {
     return (cl_mem)gpuDeviceInterface::get_gpu_memory_array(name, index, len);
 }
 cl_context& clDeviceInterface::get_context() {
